+++ Scripts/PlayerInputActionRelay.cs
@@ -4,17 +4,20 @@ using UnityEngine.InputSystem;
 
 namespace DevPeixoto.InputSystemUtils.InputSystemActionRelay
 {
-    [RequireComponent(typeof(PlayerInput))]
     [AddComponentMenu("DevPeixoto/Input System Utils/Player Input Action Relay")]
     public class PlayerInputActionRelay : MonoBehaviour
     {
-        public List<PlayerInputActionGroup> inputActionGroups = new List<PlayerInputActionGroup>();
-        
-        PlayerInput playerInput;
+        [SerializeField] PlayerInput playerInput;
+
+        public List<PlayerInputActionGroup> inputActionGroups = new List<PlayerInputActionGroup>();        
 
         private void Awake()
         {
-            playerInput = GetComponent<PlayerInput>();
+            if (playerInput == null)
+            {
+                Debug.LogError("No Player Input set");
+            }
+
             foreach (var group in inputActionGroups)
             {
                 group.inputAction = playerInput.actions.FindAction(group.InputActionRef.action.id);
@@ -23,6 +26,9 @@ namespace DevPeixoto.InputSystemUtils.InputSystemActionRelay
 
         void OnEnable()
         {
+            if (playerInput == null)
+                Debug.LogError("No Player Input set");
+
             foreach (var group in inputActionGroups)
             {
                 group.coroutineRunner = this;
@@ -32,6 +38,9 @@ namespace DevPeixoto.InputSystemUtils.InputSystemActionRelay
 
         void OnDisable()
         {
+            if (playerInput == null)
+                Debug.LogError("No Player Input set");
+
             foreach (var group in inputActionGroups)
             {
                 group.OnDisable();
@@ -41,6 +50,9 @@ namespace DevPeixoto.InputSystemUtils.InputSystemActionRelay
 #if UNITY_EDITOR
         private void OnValidate()
         {
+            if (playerInput == null)
+                Debug.LogError("No Player Input set");
+
             foreach (var group in inputActionGroups)
             {
                 group.OnValidate();